Add bullseye and make python optional for GDB

+++ scripts/build.sh
@@ -24,6 +24,8 @@ JOBS=$(($(nproc) * 2))
 codename="$(lsb_release -c -s)"
 if [ "$codename" == "buster" ]; then
 	GLIBC_VERSION=2.28 # Buster
+elif [ "$codename" == "bullseye" ]; then
+	GLIBC_VERSION=2.31 # Bullseye
 elif [ "$codename" == "bookworm" ]; then
 	GLIBC_VERSION=2.36 # Bookworm
 elif [ "$codename" == "jammy" ]; then
@@ -350,12 +352,13 @@ build_toolchain() {
 	build_package expat "$SRC_DIR/expat-${EXPAT_VERSION}" "$BINUTILS_OPTIONS"
 	GDB_OPTIONS="$BINUTILS_OPTIONS"
 	if [ "$1" = "$(gcc -dumpmachine)" ]; then
-		GDB_OPTIONS="$GDB_OPTIONS --with-python --with-system-gdbinit-dir=$WRK_DIR/$1_$2-${GCC_VERSION}/share/gdb/system-gdbinit"
+		GDB_OPTIONS="$GDB_OPTIONS $GDB_WITH_PYTHON --with-system-gdbinit-dir=$WRK_DIR/$1_$2-${GCC_VERSION}/share/gdb/system-gdbinit"
 	fi
 	build_package gdb "$SRC_DIR/gdb-${GDB_VERSION}" "$GDB_OPTIONS"
-	rm -f $PREFIX/share/gdb/system-gdbinit/wrs-linux.py
-	rm -f $PREFIX/share/gdb/system-gdbinit/elinos.py
-	cat <<'EOF' > $PREFIX/share/gdb/system-gdbinit/libstdc++.py
+	if [[ -n "$GDB_WITH_PYTHON" ]]; then
+		rm -f $PREFIX/share/gdb/system-gdbinit/wrs-linux.py
+		rm -f $PREFIX/share/gdb/system-gdbinit/elinos.py
+		cat <<'EOF' > $PREFIX/share/gdb/system-gdbinit/libstdc++.py
 import glob
 import os
 import sys
@@ -364,6 +367,7 @@ libstdcpp = os.path.abspath(glob.glob(os.path.join(os.path.dirname(__file__), '.
 sys.path.append(gcc_python_folder)
 exec(open(libstdcpp).read())
 EOF
+	fi
 	unset LOADLIBES
 	rm -rf "$BUILD_DIR"
 	rm -rf "$SYSROOT"
